Update: Move search utils to client
