+++ packages/client/.babelrc-cjs.cjs
@@ -1,4 +1,14 @@
 module.exports = {
-  presets: ['@babel/preset-typescript', '@babel/preset-env'],
+  presets: [
+    '@babel/preset-typescript',
+    [
+      '@babel/preset-env',
+      {
+        modules: 'commonjs',
+        useBuiltIns: 'usage',
+        corejs: '3.22',
+      },
+    ],
+  ],
   plugins: ['import-graphql'],
 };
