+++ packages/client/.browserslistrc
@@ -1,0 +1,5 @@
+# Browsers that we support
+
+defaults
+not IE 11
+maintained node versions
