+++ packages/client/package.json
@@ -61,6 +61,7 @@
     "test": "npm-run-all --aggregate-output --silent -n -p -c test:*"
   },
   "dependencies": {
+    "core-js": "^3.22.8",
     "lodash": "^4.17.21",
     "moment": "^2.29.3",
     "tiny-invariant": "^1.2.0"
