+++ packages/client/tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "target": "ES2015",
     "module": "ES2020",
+    "lib": ["ESNext"],
 
     "strict": true,
     "noUnusedLocals": true,
