--- packages/data/app/utils/search.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-data/utils/search';
